requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.3.0
email-validator==1.3.1
environs==9.5.0
fastapi==0.92.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
orjson==3.8.6
packaging==23.0
pydantic==1.10.4
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-decouple==3.7
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
